Handle window resize
I don't actually know if this makes the GPU destroy the previous texture data
and replace it with the new one. I initially had code to unbind, delete,
regenerate and rebind a new texture id (which worked) but when I disabled that
it still worked. There is no card-agnostic way to query GPU memory usage so I
just have to hope.