Very basic GCode implementation
This is actually a slightly difficult problem. It would be better if the
"gcode" command was completely unaware of lifting or lowering the Z axis,
because it's not really its domain. However, the tool must move to the starting
point of the LineString and then lower the tool to the cutting depth before
following the LineString. The lowering of the tool needs to be "injected" after
the move-to LineString[0] point and before the cut-to LineString[1] point.

Maybe it would be better to have "beforep0" and "afterp0" arguments where the
user can provide literal GCode for injection. Then if the user wants to add say
coolant-start right before cutting they can just do that. Not sure yet.