Add "reverse" command for reversing direction of a LINESTRING
I noticed that "moon.svg" looked messed up when generating a toolpath:

    cat resources/moon.svg \
        | ./svg2wkt \
        | ./cnc translate --dx=30 --dy=30 \
        | ./cnc toolpath --echo \
        | ./cnc-view2d

There is a bug in `toolpath` when there are "complex" corner geometries like
this, but it also showed me that OpenSCAD was exporting SVG paths with a
clockwise winding. This can be corrected using the "reverse" command:

    cat resources/moon.svg \
        | ./svg2wkt \
        | ./cnc reverse \
        | ./cnc translate --dx=30 --dy=30 \
        | ./cnc toolpath --echo \
        | ./cnc-view2d

But *beware* - cutting out a hole shaped like the "moon" is still buggy.