Broken arrows
My attempt to get arrows working. I want to show the direction of lines around
a path. It works for the simple cube, but a more complex shape makes the
program crash. It's also stupid complex for just drawing some freakn arrows. It
might be better to make an arrow triangle and then use matrix transformations
to move them into place on the lines.