Polyfill implementation
It seems to mostly work. But there are a few things still remaining:

1) The lines go right up to the top and bottom edges, rather than be backed off
a bit. This might result in unsightly tool marks on the inner edge.
2) For sharp internal corners the X backoff probably wont work properly.
3) For rotated objects the starting line looks a bit off. For example:

    echo "LINESTRING(0 0, 20 0, 20 20, 0 20, 0 0)" \
        | cnc rotate --angle=10 \
        | cnc translate --dx=10 --dy=10 \
        | cnc polyfill --echo \
        | cnc-view2d

See how it gets confused and generates a diagonal line? I might need to add
some logic so that the code is "aware" which intersection point is at the top
and which is at the bottom.

I wonder if a better algo might be to "rasterize" the area of the polygon into
a grid. Then throw away any square that has a line through it. Then walk the
squares like a flood-fill.