Added PolyFillNew function stub
With implmentation details. The idea is to use PointInPoly to generate a
rasterized grid of locations, each location represents a square of size 90% of
tool diameter. Then zigzag walk this grid generating tool paths. However
corners may be missed so keep scanning grid for 1s. Mark each visited square as
2 to avoid duplicates. On each direction change in zigzag append that point to
linestring. Result is a MultiLineString representing all the zigzag paths
required to cut out the poly. Then a final inner border cut should be made to
clean the edge.