Little bit different way of doing polyfill
Prior to this commit, the logic around diagonal movements had a bug: Under
certain conditions the diagonal movement could cross over a sharp inflection,
making the tool cut break out of the bounds of the polygon. This commit keeps
the diagonal movement but every time it happens a new LineString is created
which will force the gcode generator to life the tool above the work.

I also changed the way the raster is rescanned after a break in a zigzag: the
state of where we are in the raster and the `Yincr` is kept throughout the
polyfill (PolyFillVector is passed as a pointer and modified as a state). When
a new path is returned we scan for the next free raster point (value == 1)
starting from the last position and continuing in the direction of the `Yincr`.